seatResponse

+++ src/main/java/com/example/CinemaRoom/dto/SeatsResponse.java
@@ -3,5 +3,5 @@ package com.example.CinemaRoom.dto;
 import com.example.CinemaRoom.model.Seat;
 import java.util.List;
 
-public record SeatsResponse(long rows, int columns, List<Seat> seat) {}
+public record SeatsResponse(int rows, int columns, List<Seat> seat) {}
