+++ src/main/java/com/example/CinemaRoom/repository/SeatsRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SeatsRepository extends JpaRepository<Seats, Long> {}
+public interface SeatsRepository extends JpaRepository<Seats, Integer> {}
